moving cluster_eventmapper and gateway_eventmapper to an events package.
