+++ pkg/controllers/dnspolicy/dnspolicy_controller.go
@@ -38,21 +38,16 @@ import (
 
 	"github.com/Kuadrant/multicluster-gateway-controller/pkg/_internal/conditions"
 	"github.com/Kuadrant/multicluster-gateway-controller/pkg/apis/v1alpha1"
+	"github.com/Kuadrant/multicluster-gateway-controller/pkg/controllers/events"
 	"github.com/Kuadrant/multicluster-gateway-controller/pkg/controllers/gateway"
+	"github.com/Kuadrant/multicluster-gateway-controller/pkg/controllers/shared"
 	"github.com/Kuadrant/multicluster-gateway-controller/pkg/dns"
 )
 
 const (
-	DNSPolicyFinalizer           = "kuadrant.io/dns-policy"
-	DNSPoliciesBackRefAnnotation = "kuadrant.io/dnspolicies"
-	DNSPolicyBackRefAnnotation   = "kuadrant.io/dnspolicy"
+	DNSPolicyFinalizer         = "kuadrant.io/dns-policy"
+	DNSPolicyBackRefAnnotation = "kuadrant.io/dnspolicy"
 )
-
-type DNSPolicyRefsConfig struct{}
-
-func (c *DNSPolicyRefsConfig) PolicyRefsAnnotation() string {
-	return DNSPoliciesBackRefAnnotation
-}
 
 // DNSPolicyReconciler reconciles a DNSPolicy object
 type DNSPolicyReconciler struct {
@@ -151,7 +146,7 @@ func (r *DNSPolicyReconciler) reconcileResources(ctx context.Context, dnsPolicy 
 	}
 
 	// reconcile based on gateway diffs
-	gatewayDiffObj, err := r.ComputeGatewayDiffs(ctx, dnsPolicy, targetNetworkObject, &DNSPolicyRefsConfig{})
+	gatewayDiffObj, err := r.ComputeGatewayDiffs(ctx, dnsPolicy, targetNetworkObject, &shared.DNSPolicyRefsConfig{})
 	if err != nil {
 		return err
 	}
@@ -176,7 +171,7 @@ func (r *DNSPolicyReconciler) reconcileResources(ctx context.Context, dnsPolicy 
 func (r *DNSPolicyReconciler) deleteResources(ctx context.Context, dnsPolicy *v1alpha1.DNSPolicy, targetNetworkObject client.Object) error {
 	// delete based on gateway diffs
 
-	gatewayDiffObj, err := r.ComputeGatewayDiffs(ctx, dnsPolicy, targetNetworkObject, &DNSPolicyRefsConfig{})
+	gatewayDiffObj, err := r.ComputeGatewayDiffs(ctx, dnsPolicy, targetNetworkObject, &shared.DNSPolicyRefsConfig{})
 	if err != nil {
 		return err
 	}
@@ -228,16 +223,16 @@ func (r *DNSPolicyReconciler) readyCondition(targetNetworkObjectectKind string, 
 }
 
 func (r *DNSPolicyReconciler) SetupWithManager(mgr ctrl.Manager) error {
-	gatewayEventMapper := &GatewayEventMapper{
+	gatewayEventMapper := &events.GatewayEventMapper{
 		Logger: r.Logger().WithName("gatewayEventMapper"),
 	}
 	client := r.Client()
 	dnsHelper := dnsHelper{Client: client}
 	r.dnsHelper = dnsHelper
-	clusterEventMapper := &ClusterEventMapper{
+	clusterEventMapper := &events.ClusterEventMapper{
 		Logger:             r.Logger().WithName("clusterEventMapper"),
 		GatewayEventMapper: gatewayEventMapper,
-		client:             r.Client(),
+		Client:             r.Client(),
 	}
 	return ctrl.NewControllerManagedBy(mgr).
 		For(&v1alpha1.DNSPolicy{}).
