+++ pkg/controllers/shared/shared_refs.go
@@ -1,0 +1,11 @@
+package shared
+
+const (
+	DNSPoliciesBackRefAnnotation = "kuadrant.io/dnspolicies"
+)
+
+type DNSPolicyRefsConfig struct{}
+
+func (c *DNSPolicyRefsConfig) PolicyRefsAnnotation() string {
+	return DNSPoliciesBackRefAnnotation
+}
